requirements.txt
crewai
crewai[tools]
langchain-google-genai
python-dotenv
cairosvg
Pillow
aiohttp
